requirements.txt
torch==2.3.0
torchvision==0.18.0
monai==1.4.0
transformers==4.26.1
safetensors==0.4.3
salesforce-lavis==1.0.2